requirements.txt

fastapi
uvicorn[standard]
python-docx
pdfminer.six
language-tool-python
spacy
textstat
pytest
